A simple attendance marker app.  To run, clone the project and open the project in any IDE of your choice and locate the project directory in your terminal then run flutter pub get to get all the dependencies and then run flutter run.  We only worked on taking photos, the photo is only available in app, no network request and local storage yet
A simple attendance marker app.

To run, clone the project and open the project in any IDE of your choice and locate the project directory in your terminal then run flutter pub get to get all the dependencies and then run flutter run.

We only worked on taking photos, the photo is only available in app, no network request and local storage yet